code refactored to reduce code duplication
